+++ packages/core/src/modules/credentials/CredentialsApi.ts
@@ -16,6 +16,7 @@ import type {
   SendCredentialProblemReportOptions,
   DeleteCredentialOptions,
   SendRevocationNotificationOptions,
+  DeclineCredentialOfferOptions,
 } from './CredentialsApiOptions'
 import type { CredentialProtocol } from './protocol/CredentialProtocol'
 import type { CredentialFormatsFromProtocols } from './protocol/CredentialProtocolOptions'
@@ -48,7 +49,7 @@ export interface CredentialsApi<CPs extends CredentialProtocol[]> {
   // Offer Credential Methods
   offerCredential(options: OfferCredentialOptions<CPs>): Promise<CredentialExchangeRecord>
   acceptOffer(options: AcceptCredentialOfferOptions<CPs>): Promise<CredentialExchangeRecord>
-  declineOffer(credentialRecordId: string): Promise<CredentialExchangeRecord>
+  declineOffer(credentialRecordId: string, options?: DeclineCredentialOfferOptions): Promise<CredentialExchangeRecord>
   negotiateOffer(options: NegotiateCredentialOfferOptions<CPs>): Promise<CredentialExchangeRecord>
 
   // Request Credential Methods
@@ -324,12 +325,22 @@ export class CredentialsApi<CPs extends CredentialProtocol[]> implements Credent
     return credentialRecord
   }
 
-  public async declineOffer(credentialRecordId: string): Promise<CredentialExchangeRecord> {
+  public async declineOffer(
+    credentialRecordId: string,
+    options?: DeclineCredentialOfferOptions
+  ): Promise<CredentialExchangeRecord> {
     const credentialRecord = await this.getById(credentialRecordId)
     credentialRecord.assertState(CredentialState.OfferReceived)
 
     // with version we can get the Service
     const protocol = this.getProtocol(credentialRecord.protocolVersion)
+    if (options?.sendProblemReport) {
+      await this.sendProblemReport({
+        credentialRecordId,
+        description: options.problemReportDescription ?? 'Offer declined',
+      })
+    }
+
     await protocol.updateState(this.agentContext, credentialRecord, CredentialState.Declined)
 
     return credentialRecord
@@ -532,29 +543,40 @@ export class CredentialsApi<CPs extends CredentialProtocol[]> implements Credent
   /**
    * Send problem report message for a credential record
    * @param credentialRecordId The id of the credential record for which to send problem report
-   * @param message message to send
    * @returns credential record associated with the credential problem report message
    */
   public async sendProblemReport(options: SendCredentialProblemReportOptions) {
     const credentialRecord = await this.getById(options.credentialRecordId)
-    if (!credentialRecord.connectionId) {
-      throw new CredoError(`No connectionId found for credential record '${credentialRecord.id}'.`)
-    }
-    const connectionRecord = await this.connectionService.getById(this.agentContext, credentialRecord.connectionId)
 
     const protocol = this.getProtocol(credentialRecord.protocolVersion)
-    const { message } = await protocol.createProblemReport(this.agentContext, {
+
+    const offerMessage = await protocol.findOfferMessage(this.agentContext, credentialRecord.id)
+
+    const { message: problemReport } = await protocol.createProblemReport(this.agentContext, {
       description: options.description,
       credentialRecord,
     })
-    message.setThread({
-      threadId: credentialRecord.threadId,
-      parentThreadId: credentialRecord.parentThreadId,
-    })
+
+    // Use connection if present
+    const connectionRecord = credentialRecord.connectionId
+      ? await this.connectionService.getById(this.agentContext, credentialRecord.connectionId)
+      : undefined
+    connectionRecord?.assertReady()
+
+    // If there's no connection (so connection-less, we require the state to be offer received)
+    if (!connectionRecord) {
+      credentialRecord.assertState(CredentialState.OfferReceived)
+
+      if (!offerMessage) {
+        throw new CredoError(`No offer message found for credential record with id '${credentialRecord.id}'`)
+      }
+    }
+
     const outboundMessageContext = await getOutboundMessageContext(this.agentContext, {
-      message,
-      associatedRecord: credentialRecord,
+      message: problemReport,
       connectionRecord,
+      associatedRecord: credentialRecord,
+      lastReceivedMessage: offerMessage ?? undefined,
     })
     await this.messageSender.sendMessage(outboundMessageContext)