+++ packages/core/src/modules/vc/models/credential/__tests__/W3cCredential.test.ts
@@ -152,7 +152,7 @@ describe('W3cCredential', () => {
         },
         W3cCredential
       )
-    ).toThrow()
+    ).not.toThrow()
 
     expect(() =>
       JsonTransformer.fromJSON(